fix comments

+++ WeatherAPI/GismetioClient/Contracts/SpeedWindContract.cs
@@ -5,6 +5,6 @@ namespace GismetioClient.Contracts
     public class SpeedWindContract
     {
         [JsonProperty("m_s")]
-        public int WindSpeed {  get; set; }
+        public float WindSpeed {  get; set; }
     }
 }
